requirements.txt
scrapy-zyte-api==0.12.2